works with 16 bits (ish)

+++ Brass2Go.X/timer.h
@@ -2,6 +2,6 @@
 #define B2G_TIMER_H
 
 
-int timer_Init(int rate);
+int timer_Init(long rate);
 
 #endif
